requirements.txt
fastapi==0.111.0
langchain==0.1.13
python-dotenv==1.0.1
langchain-openai==0.1.1
